fix proj tests
